feat(epg): add epg timelin compontent
