+++ src/components/EpgTimeline/EpgTimeline.module.scss
@@ -1,0 +1,54 @@
+@use '../../styles/variables';
+@use '../../styles/theme';
+
+.timelineBox {
+  position: relative;
+  font-size: 14px;
+}
+
+.timelineTime {
+  position: absolute;
+  top: 18px;
+  left: -18px;
+  color: variables.$white;
+}
+
+.timelineContainer {
+  position: sticky;
+  top: 0;
+  display: flex;
+  height: 60px;
+}
+
+.timelineDividers {
+  display: grid;
+  align-items: end;
+  width: 100%;
+  height: 100%;
+  padding-bottom: 6px;
+  grid-template-columns: repeat(4, 1fr);
+}
+
+.timelineDivider {
+  width: 1px;
+  height: 10px;
+  margin-right: 10px;
+  background: variables.$white;
+}
+
+.leftControl {
+  position: sticky;
+  top: 25px;
+  left: 25px;
+  width: 20px;
+  height: 25px;
+}
+
+.rightControl {
+  position: sticky;
+  top: 25px;
+  right: 25px;
+  width: 20px;
+  height: 25px;
+  background: inherit;
+}
